requirements.txt
yt-dlp>=2024.1.0
faster-whisper>=0.10.0
openai>=1.0.0
google-generativeai>=0.3.0
typer>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0
webvtt-py>=0.4.6
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.36.0
